requirements.txt
python-telegram-bot==20.7
requests==2.31.0
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
urllib3==2.1.0
numpy==1.24.3
aiohttp==3.9.5
cachetools==5.3.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
